Renamed to pyFMS and cross referenced unit tests for field_manager with FMS versions
