+++ pyFMS/.gitignore
@@ -56,6 +56,7 @@ coverage.xml
 *.py,cover
 .hypothesis/
 .pytest_cache/
+tests/__pycache__/*
 
 # Translations
 *.mo
